requirements.txt
# requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.12
pandas==2.2.3
openai==1.30.0
httpx==0.27.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-core==0.3.0
python-dotenv==1.0.1
aiofiles==24.1.0
requests==2.32.3